+++ Modern JavaScript From The Beginning/Section 7: Asynchronous JavaScript, Ajax & Fetch API/Custom HTTP Library/app.js
@@ -1,0 +1,12 @@
+const http = new HTTPLib();
+
+http.get(
+    'https://jsonplaceholder.typicode.com/posts',
+    function (error, responseText) {
+        if (error) {
+            console.log(error);
+        } else {
+            console.log(responseText);
+        }
+    }
+);